Break help into topics, add comparitive testing
This change breaks up the help output into a compact general usage
output, and adds the following sections as help topics: examples,
formatCodes, timeSyntax, and timeZone.

Tests updated to exercise the new help topics and error-handling.

Added comparative testing to the test tool. This allows you to capture
the output of a given command, and then test that the output of some
other command yields the same result. For one, this avoids long accepted
output with the same help text over and over. It will also provide a way
later to test output that will change from run to run, such as file
creation, modification and access times.

Fixes #19, #21