Switch to wchar_t, wstring, etc
Unicode on Windows is a total mess.

Output to the console does not work as expected, though it sort of
changes depending on the console code page (via the `chcp` command). The
`SetConsoleOutputCP()` function doesn't appear to have any effect from
within the program.

In addition, constructing UTF-8 sequences via `\xfff` escapes doesn't
appear to work, as it's fragile when emitting a code unit
(partially-encoded code points) a byte at a time. This kills the option
of supporting `\xfff` escape sequences as an option.

In general, the whole things a mess.

What's working right now is that everything's now using UTF-16 as per
Microsoft recommendations. Depending on your codepage, you can enter
some characters on the command line and have them printed back out when
redirecting to a file or piping the output. It doesn't work when going
to the console output. Unpredictable and unsatisfying, but I've decided
to keep the conversion work to wide chars just in case.

Fixes #9