Fix crashing bug in array count argument.
wcsftime() takes an array and the number of elements of that array. On
converting from char to wchar_t, this introduced the bug that a
1024-element destination buffer was passed as size 2048, when it should
have been 1024.

Apparently wcsftime() pads out the remainder destination buffer, and so
we got buffer overflow.

It might be better to use std::array in the future.